+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * change: Toastx.show\*(\*, String) method change to show\*(\*, CharSequence)
 * fix: Fix the bug that Toastx.show\*(\*, int) method may not get multilingual information
 * fix: Fix the bug that there may be null elements in the array returned by Storagex.getAppExternal\*Dirs() method
+* new: Storagex.clean\*Dirs() method now returns boolean
 
 ## v1.0.0-alpha09
 * remove: Remove dependency on javax